+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [3.2.0] Aug-19-2019
 
 ### Additions
-
+- [Issue #189] Add CanReadAll() and CanWriteAll()
 
 ## [3.1.0] Aug-06-2019